TelegramBot hierarchy fix
